requirements.txt
matplotlib>=2.2.2
cairocffi>=0.9.0
pkgconfig>=1.4.0
setuptools>=39.1.0
scipy>=1.1.0
numpy>=1.15.4
opencv-python
absl-py==1.0.0
gym==0.21.0
tensorflow-probability==0.16.0
tf-agents[reverb]==0.13.0
torch==1.6.0